requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
PyYAML==6.0.3
safetensors==0.6.2
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
Werkzeug==3.1.3
